Fix xtd::net::sockets::ip_v6_multicast_option::interfce_index to uint64

+++ src/xtd.core.native.linux/src/xtd/native/linux/socket.cpp
@@ -183,10 +183,10 @@ int32_t socket::get_socket_multicast_option(intmax_t handle, int32_t socket_opti
   return result;
 }
 
-int32_t socket::get_socket_ip_v6_multicast_option(intmax_t handle, int32_t socket_option_name, vector<uint8_t>& multicast_address, uint32_t& interface_index) {
+int32_t socket::get_socket_ip_v6_multicast_option(intmax_t handle, int32_t socket_option_name, vector<uint8_t>& multicast_address, uint64_t& interface_index) {
   struct multicast {
     uint8_t multicast_address[16];
-    uint32_t interface_index;
+    uint64_t interface_index;
   } m;
   size_t multicast_size = 0;
   auto result = getsockopt(static_cast<int32_t>(handle), IPPROTO_IP, socket_option_name_to_native(socket_option_name), &m, reinterpret_cast<socklen_t*>(&multicast_size));
@@ -332,10 +332,10 @@ int32_t socket::set_socket_multicast_option(intmax_t handle, int32_t socket_opti
   return setsockopt(static_cast<int32_t>(handle), IPPROTO_TCP, socket_option_name_to_native(socket_option_name), &m, static_cast<socklen_t>(sizeof(multicast)));
 }
 
-int32_t socket::set_socket_ip_v6_multicast_option(intmax_t handle, int32_t socket_option_name, const vector<uint8_t>& multicast_address, uint32_t interface_index) {
+int32_t socket::set_socket_ip_v6_multicast_option(intmax_t handle, int32_t socket_option_name, const vector<uint8_t>& multicast_address, uint64_t interface_index) {
   struct multicast {
     uint8_t multicast_address[16];
-    uint32_t interface_index;
+    uint64_t interface_index;
   } m;
   for (auto index = 0U; index < multicast_address.size(); ++index)
     m.multicast_address[index] = multicast_address[index];
